Update and rename M05_project.py to main.py
refactor: main
